Added Export Functionnalite

+++ src/main/java/com/example/tourisme_medicale/models/Chirurgie.java
@@ -3,12 +3,12 @@ public class Chirurgie extends Type {
 
     private int id;
     private float prix;
-    private float duree;
+    private int duree;
     private Specialite specialite;
 
     private String typeChirurgie;
 
-    public Chirurgie(int id, float prix, Specialite specialite,  float duree, String typeChirurgie) {
+    public Chirurgie(int id, float prix, Specialite specialite,  int duree, String typeChirurgie) {
         super(id, prix, specialite);
         this.id = id;
         this.prix = prix;
@@ -18,11 +18,11 @@ public class Chirurgie extends Type {
     }
 
 
-    public float getDuree() {
+    public int getDuree() {
         return duree;
     }
 
-    public void setDuree(float duree) {
+    public void setDuree(int duree) {
         this.duree = duree;
     }
